Revert "Refactor generate secret task" temporarily
This reverts commit 1986849a2a9c0315943a7ac19c54731e164c8d4e.

Although this change made the task easier to understand, using `stdin:`
to feed ansible-vault sometimes caused "broken pipe" errors when a shell
exited without completely consuming stdin.

Unfortunately, we can't sidestep the use of flock(1) by using run_once,
see f6bd528 for details. The better solution would be to write a module
that uses module_utils.common.file.FileLock to implement the locking in
Python, without this complex sh+flock invocation.

However, since this problem was reported by tpaexec users and not found
during development, it seems safest to revert the recent change until we
can either write such a module (which is not a priority), or work around
the problem with using `stdin:` in another way (which is in the works).