Simplify generate secret
Use ansible native features for constructing dictionary, dumping yaml
and calling the password script.
Improve readability in main command.
Make sure non-zero exit status from ansible-vault is visible.

This reinstates the changes reverted by 0741045, with the use of <<< to
avoid the problems with using `stdin:` to feed ansible-vault (which did
not consume stdin in the cases where flock failed to acquire the lock).