Allow for Architecture class usage in tests (#274) (#336)
The Architecture class can be instantiated now outside the CLI
interface by supplying static paths to the architecture directory, the
tpaexec lib directory and a fixed argument list for argparse.

Determination of the path vars for architecture configure entrypoint
scripts has been normalised to use a single configure function that
interrogates filesystem paths via the os module.
A factory class can is used to create a class instance based on the name
of the architecture needed.
Both these allow for the configure boilerplate to be implemented once.

A simple proof-of-concept test has been provided to show how this can be
used. This will be extended in future commits. The test assert
statements are not important at this point. They are just there to
demonstrate what could be done.