Set cidr by default in ec2_vpc (#444)
For --platform aws, we used to generate the following clause by default

    ec2_vpc:
        Name: Test

This depends on a VPC named Test (with CIDR 10.33.0.0/16) being present,
which it was in every region under the 2ndQuadrant AWS account. However,
it would fail if the VPC was not present, and the user would have to add
a CIDR to create a new VPC, or change the name to use an existing one.

With the majority of users now using non-2ndQuadrant AWS accounts, this
is quite unfriendly, and requires extra warnings in the documentation.
The rationale for not creating the VPC by default was the total limit of
5 VPCs and people not having permission to create more, which is not as
relevant now when most people use isolated accounts, not just individual
IAM users under the same AWS account.

So we now specify a CIDR by default, so that the VPC will be created if
it doesn't exist (or just used, if it does).