Bump to openjdk 11 for efm on distro providing it
Change the way we define per distro version package to manage distro
version and keep java 8 when distro doesn't provide 11.
This is the curated way of handling multiple distro and versions
https://ansible-tips-and-tricks.readthedocs.io/en/latest/os-dependent-tasks/variables/
The change was only applied to dependency package since it would require
to change packages_for() filter.

We use `ansible_distribution` for Ubuntu since it outputs `Debian`
as `ansible_os_family`

Closes #463