Perform input data transformations for instance attributes
In instance.volumes we need to support the volume_type 'none' in other
cloud providers. This is achieved by extracting the relevant step from
the previously aws specific filter expand_instance_volumes.

In instance.location we support the input value to be either a string
or an int (for backwards compatibility). This is coerced to a string
so that it can be uniform during runtime and location.Name can be used
to index the formatting of data for subnets going into terraform.