TPA-184: Add command to rebuild sources
At deploy-time, we install onto each host scripts to rebuild anything
which was build from source on that host. A new command,
rebuild-sources, runs the scripts without invoking any of the usual
deploy playbooks, thus enabling a much faster rebuild cycle.

Rebuild scripts are dropped into /etc/tpa/rebuild-scripts and the
rebuild-sources command invokes /etc/tpa/rebuild-sources.sh, which runs
all the scripts in that directory and then restarts postgres if
appropriate.

If the sources are local to the cluster machines, we run "git pull"
before rebuilding.

The rebuild-sources command is documented in the main documentation for
source builds, and in the command-line help.