Add filter to convert a list of maps to an indexed map
All variables that are lists of dicts that terraform parses are
converted in ansbile to dicts with a unique key name. A custom filter
has been added to perform this action in ansible.
Although, it is possible to do this in ansible natively with loops when
a `dict.Name` key is needed, by using `loop_control` and `index_var`
when the key name is the index, a reasonable solution for nested lists,
like with `instances.volumes` is not clear.