psycopg error with postgres_version < 10
Postgres 9.6 repositories are not available through the rpm package
published on PGDG repositories. we need to be able to support this
version for now.
- add check for postgres version on tasks that try to modify the repo
file and would pollute the file since there is no definition for pg<=9.6
- add the possibility to define `exclude` section when adding repo
to the list in config.yml
- add version check when trying to export PGHOST env variable to
avoid comma separated list which are not supported by pg < 10

For end users, the solution is to manually add the repository to
the list in config.yml file:
```
yum_repositories or apt_repositories depending on the OS family in use:
yum_repository_list: [EPEL, PGDG, PGDG96, PGDG96-DEBUGINFO]
  yum_repositories:
    PGDG96:
      description: "Postgres 9.6 Repo"
      file: 'pgdg96'
      baseurl: 'https://download.postgresql.org/pub/repos/yum/9.6/redhat/rhel-$releasever-$basearch'
      gpgkey: 'file:///etc/pki/rpm-gpg/RPM-GPG-KEY-PGDG'
      exclude: 'python-psycopg2 python2-psycopg2 python3-psycopg2'
    PGDG96-DEBUGINFO:
      description: "Postgres 9.6 Debug Info Repo"
      file: 'pgdg96-debuginfo'
      baseurl: 'https://download.postgresql.org/pub/repos/yum/debug/9.6/redhat/rhel-$releasever-$basearch'
      gpgkey: 'file:///etc/pki/rpm-gpg/RPM-GPG-KEY-PGDG'
```
exclude list is mandatory for this repository definition.
Closes #176