Remove Boto dependency and switch to Boto3
* Use iam_role module instead of iam
The iam module uses the deprecated boto library, which is unmaintained
and does not support SSO/temporary credentials. If you export an access
key id, access key, and security token in the environment, or spell them
out in ~/.aws/credentials, things work, but if you use `aws sso login`
and a profile name without explicitly setting credentials, boto can't
find credentials to use (but actually complains about not being able
to find the profile, even if it's in ~/.aws/config).
The iam_role module, which uses botocore, will suffice for our purposes.
It requires the caller to specify an `assume_role_policy_document`, in
which we spell out that we only want to allow EC2 instances to be able
to use this role as their instance profile. (The policy is the same as
what the role ended up with when using the old module.)

* Change Module from ec2 to ec2_instance
-  change volume attributes to match the expected format for new
ec2_instance module. the new module won't work if any unexpected
key is present.
- output format and parameters are differents too and we use them
through the run, we needed to adapt all the tasks that use the output of
ec2_instance module.
- ec2_instance outputs doesn't filter instances using `Name` tag
unless it's `name` paramter is used, this caused issues when running
provision several times (instances in the same subnet would all be
renamed with the same name.
- added some label for tasks using loops to make the output readable.

* Inventory change from script to plugin
- changed ansible.cfg file to use aws_ec2 plugin
- added aws_ec2.yml config file template
- changed ec2_instance to instance_id
- changed ec2_region to placement.region
- changed the source to set up name from ec2_tag_Name to tags['Name']

* Add keyed_groups
add multiple keyed_groups entries to match groups created by previously
used script.

* Change condition as list
avoid having tests on streaming_from when this variables is undefined

* Add cache configuration for aws_ec2 plugin
- enable cache
- set jsonfile as cache plugin
- set timeout
- prefix cache with cluster name
- store cache in tmp folder.

* Add filter to compute expected slots
added a filter taking list of replica dict and the hostname of the
current node to output the list of slot_name expected by the system.
I chose to use list of replica instead of hostvars to simplify the code
for the filter since the construction of this list is straightforward
using jinja filters.

* Ansible linting

* Cleaning unused boto related patch.
- Remove the patch we apply to a boto file during setup since boto is not
installed anymore.
- Remove unused ec2.py file since we will be using the inventory plugin
from aws collection.
- Remove unused variable.
- With boto to boto3 switch we don't need to patch boto file anymore.

* Use community.aws and amazon.aws collections
- Install community.aws collection which is dependant on amazon.aws
collection.
- Change module calls to specify the collection name and ensure we use
the installed collection.
adapt module inputs variables to newer version of modules.

* Remove boto as dependancy
We don't need boto as a dependancy anymore.

* Change expected state of ec2_instance creation
Change ec2_instance module expected state for instances during creation
from `present` (default = ensures instances exist, but does not
guarantee any state) to `started` (= `present` + ensures the instances
are running + waits for EC2 status checks to report OK if wait=true)
in order to avoid reaching latter tasks without network information
available.

* In a deprovision the instance was not removed due to a problem detaching
a disk. The deprovision was repeated and failed when attempting to
remove an attached policy because the policy had been removed in the
previous run.
This change ensures the operation only occurs if the IAM role exists.
* Add cleaning steps if leftover files from ec2.py
we need to cleanup old inventory script and config file if those exist.