check bdr_version instead of postgres_version
for 9.4 we only support bdr use of version 1. we don't
need to modify the config file for stock postgres9.4 since this change
won't make the deployment work due to package version issues not
resolved by this change.
we use bdr_version instead of postgres_flavour because flavour can be
'None' at this point we set it later to 'postgresql-bdr' if
bdr_version == 1.