Simplify _install_pip_packages
We always run pip with `-r ansible.txt -r requirements.txt`, but the
locations of those files can vary depending on whether ANSIBLE_HOME is
set or tpaexec-deps is installed. e20e365 simplified this logic, but we
can do better still by introducing {ANSIBLE,TPA}_REQUIREMENTS, which are
set to $TPA_DIR/*.txt by default and changed in the ANSIBLE_HOME/PIP_DIR
branches if needed.

Using an associative array to declare contrib files to copy also helps
to reduce the likelihood of bugs like the one fixed in b285340.