Remove lxd support
Although lxd itself is still a viable, maintained platform (5.0 LTS was
just released in April 2022), and has some advantages over using Docker,
the support for lxd in tpaexec is unused and unmaintained.

Since there hasn't been any perceptible interest in using it for a long
time, we're better off removing the support until that changes.

Closes #327