Remove BDR-Simple architecture
This architecture was created for experimentation with BDR, and is not
something we want to encourage customers to use (even for testing). It
might still be useful for internal testing/prototyping, but since the
functionality is available (and existing clusters won't be affected),
we can accommodate that use-case even without publishing a separate,
user-selectable architecture.

(We have some of the infrastructure required to package architectures
separately from tpaexec, but it doesn't seem worth doing in this case.)

Closes TPA-10