Restart etcd processes one by one
When an etcd cluster is already serving as the consensus layer for HARP,
we can't afford to restart etcd on all instances simultaneously, because
that will cause HARP to become unhappy and break routing.

So in this change we:

1. Start or restart etcd only one instance at a time within
   groups['role_etcd']

2. Before restarting, we run `etcd endpoint health --endpoints …` to
   ensure the cluster is OK (and fail otherwise), then restart, then
   check endpoint health again (but be prepared to retry a few times
   until everything is back to normal (in my tests it was always fine
   the first time)) before moving on to the next etcd.

Now that we make sure the etcd cluster reports that it's fine before and
after, restarting any single etcd should not bother either etcd or harp.

References: TPA-8, RT80790