Clean up stray 2ndQuadrant references
I haven't touched the "2ndQuadrant repository" references, because they
are still in use, and calling them "EDB repositories" next to where we
describe support for the other "EDB repositories" is going to confuse
people.

I haven't touched the one in architecture-BDR-Simple.md because there's
a PR pending to remove the architecture, and I don't want conflicts.