Lessen maintenance on mkdocs requirements
We do not depend on mkdocs in this repo as it is build downstream in the
tpaexec-packaging jenkins job. That uses docs/Dockerfile.mkdocs which
contains the deps for mkdocs. There is no version pinning in this file.
Therefore, matching the packages it installs.

By moving the requirements.in to requirements.txt we will no longer get
pip-compiled requirements.txt, as it serves no purpose as above and
ensure dependabot does not issue suggested upgrades, therefore lessening
the overhead of maintenance whilst allowing ad-hoc doc builds via make.