Change default_harp_consensus_protocol to bdr for harp_version 2
Given the problems we've encountered with using etcd for harp consensus
(one of which we fixed recently by enabling compaction), we change the
default consensus protocol to BDR, which we generally prefer anyway.

To ensure that this change doesn't affect existing clusters (which are
running etcd with harp_consensus_protocol not explicitly set), we read
the DCS driver value from harp/config.yml and use that to set
harp_consensus_protocol.

We also add a complex assert to ensure that all instances (both bdr and
harp-proxy) with the same bdr_node_group value also have the same value
for harp_consensus_protocol too. (Note that different bdr_node_groups
may have different consensus settings.)

(We have to set first_bdr_primary for harp-proxy instances as part of
this change, because we're deriving the preferred value for the entire
group plus proxies from first_bdr_primary's value.)

Committed with some changes by AMS.