Ensure we're running Docker CE v20.10+
v20.10 is the first version to support the unified cgroups2 hierarchy,
which solves various problems people have reported with provisioning
containers through tpaexec on systems with recent kernels.

We don't support (and have never supported) older distribution-packaged
versions of Docker, there are too many problems with them for it to be
practical.