Ensure we're running Docker CE v20.10+ and CgroupVersion 1
v20.10 is the first version to support the unified cgroups2 hierarchy,
which solves various problems people have reported with provisioning
containers through tpaexec on systems with recent kernels. However,
support for CgroupVersion 2 is not fully baked yet for docker sdk in
ansible and related tooling. So while we recommend using a recent
version of docker; we rely on CgroupVersion 1 until version 2 is
fully supported.

We don't support (and have never supported) older distribution-packaged
versions of Docker, there are too many problems with them for it to be
practical.

Signed-off-by: Haroon <muhammad.haroon@enterprisedb.com>