Don't use openvpn by default in M1
The default M1 configuration uses two different availability zones in
the same AWS region as its two locations anyway, so it doesn't need to
use openvpn to get the instances to speak to each other. Here we change
the default on the basis that installing openvpn is an extra step most
users won't really need.

Existing clusters that use openvpn are unaffected, and new clusters can
use openvpn if required by just setting vpn_network under cluster_vars
and adding 'openvpn-server' to the role of a suitable instance.