TPA-37 replace print errors with custom exception classes
# Adds a sys exception hook to catch all unhandled exceptions.

When an exception is raised either a message is displayed from custom
exception classes or print a message in a similar style to normal. This
could be tweaked to be a more generic message allowing fpr a move all
`print->sys.exit` errors to raise custom exceptions.

# Add Exception classes types for error reporting

Exception types define a categories of fault conditions that can be used
to raise messages back up the stack to be caught by calling
applications. Each type has a MSG constant that can be called to display
a more friendly message to the user. When we want to raise a fault, pass
the specific condition that caused the error as an argument to the
class. Therefore, we get a message which states the type of error and
the specific condition which caused it. In the initial change written in
this commit, the type of error is a Platform error and the condition is
an unknown platform string discovered by the load method.

# Unhandled exceptions

All exceptions not handled in the code or raised by 3rd party libraries
display a standard error message and sent to the original sys exception
hook. A full traceback will be displayed to the user.