Add support for named arguments in postgresql_query
Within the list of queries, you can now do either of the following:

    - text: … %s, %s …
      args:
      - hello
      - world
      - "{{ optional_something|default(omit) }}" # Turned into a NULL

    - text: … %(one)s, %(two)s …
      named_args:
        one: hello
        two: world
        six: "{{ something_else|default(omit) }}" # Just doesn't work

This makes it possible to use either %s or %(name)s placeholders for a
particular query. If specified, named_args takes precedence. This form
allows the same argument to be used multiple times in the query without
repetition of the value.

If you want to pass in explicit NULLs to the query, you can't do it
using named_args and omit, because the omit keys are just not passed in
to the module. Unfortunately, using jinja2_native also does not work, so
there's no obvious way to make NULLs work with %(named) placeholders.