Require --harp-consensus-protocol for BDR-Always-ON
The choice of consensus protocol has serious effects on performance, so
it's not a decision that can be made lightly, i.e., having the default
value of etcd (or bdr) comes with so many potential consequences that
it's better to make it explicit.

It's not possible for TPAexec to set a good default anyway, because what
we want is to "use BDR unless you are deploying across high latency
network between locations in which case use per location etcd, in case
of high latency network within same location, don't use HARP at all"
(quoting Petr).

Note that this affects the configuration of newly-generated clusters,
but will not affect existing clusters that use etcd without setting
harp_consensus_protocol explicitly.