Add new pkg/{add_to_list,install} API with README
These roles define a new interface for xxx/pkg roles to follow. See
README for details, but the highlights are:

- pkg roles have a list-packages.yml that anyone can call to get a
  complete list of packages required by that role (on a particular
  instance, based on its role and other conditions)

- roles can use the same interface to include packages from other roles
  if required (e.g., postgres client packages)

- All the xxx/pkg/tasks/main files just call list-packages to get the
  list of packages, and use pkg/install to install them.

Clearly separating package enumeration from package installation makes
it possible to do other things with the package list, such as download
them for later installation.

(This commit just introduces the new interface, it doesn't change any of
the existing pkg roles to use it.)

Based loosely on a proposal from Matt.