Add `tpaexec download-packages` command
First, we introduce two additional pieces to the pkg API:

1. pkg/list_all_packages discovers all pkg roles and calls their
   list-packages.yml to assemble a complete list of packages for an
   instance

   (We add a stub list-packages.yml to zabbix_agent/pkg, which does not
   quite follow the conventions of the other pkg roles. That's fine, we
   can sort it out later.)

2. pkg/download takes a list of packages (as pkg/install does) and
   downloads them to a given download directory.

Then download-packages.yml spins up a temporary Docker container within
the cluster (using add_host, without modifying the inventory), with the
roles of all other instances in the cluster, and uses pkg/download on
the result of pkg/list_all_packages.

After downloading the packages and all dependencies, we generate apt/yum
repository metadata for the packages. (One could argue that it should be
moved out of pkg/download, which we can do in future if needed.)

We also delay the creation of any inventory directories until write.yml
is called (which the downloader doesn't do) to create the inventory at
the end of provisioning. Otherwise, if we run download-packages before
provision, we create empty directories that would be seen as having a
broken inventory on subsequent playbook runs.

(download-packages is based on a proposal by AMS, with extensive work by
Matt; and pkg/download is extracted from Matt's code.)