Avoid running sys/bootstrap if /etc/tpa exists
Although we create /etc/tpa a little after sys/bootstrap runs, we can
safely assume that the bootstrapping was completed successfully if we
find that the directory exists.

This allows us to skip a bunch of repetitive tasks early in the deploy,
which saves quite a bit of time even though the the tasks were guarded
by tests to make them return quickly if no installation was needed.