Sidestep some tasks that the downloader cannot evaluate
When running the downloader, we typically won't set first_bdr_primary,
which breaks some assumptions in roles/facts. Here we carefully extract
those tasks into a separate include that the downloader skips, because
it doesn't need those tasks anyway (though it does need other tasks in
facts to do the repository setup).

The order of tasks in roles/facts/tasks/postgres is important, and this
commit is meant to change things as little as possible.