TPA-74 Move pgdcli role call to deploy.yml
BDR 4.1 packages are now generally available we don't need to look up
the minor version of the running cluster, so can move the package
install to the correct stage in deploy.
Also changing the version comparison to major version, as pgdcli can be
installed and if ran will exit if not supported (e.g. on BDR <4.1)