Don't install repmgr on bdr instances
We use repmgr to clone bdr+replica instances (because we're set up to
use it already, so it's convenient), but (a) it's not a safe way to set
up a BDR physical replica, because we don't set up standby slot names;
and (b) repmgr is, unfortunately, essentially unmaintained, and also we
don't want to recommend expanding the scope of its use (which, in this
case, would mean publishing the packages to yet another repository).

There are no longer any recommended BDR architectures that include
physical replicas, and in the absence of a pg_basebackup-based clone
implementation (which would be good to have for various reasons), this
is the easiest way to fail if someone tries to create one.

Existing bdr+replica instances are unaffected by this change, but new
physical replicas in BDR clusters won't be able to be created.

(Note: this question came up because repmgr packages built against
PGE/2ndQPostgres are published to the BDR 3.x repositories, but not to
the BDR 4.x ones, so the downloader failed when it tried to download
repmgr. And we'd rather disable physical replica functionality than
publish the repmgr packages to yet another repository.)