Different etcd placement for the BDR-Always-ON template
When using 'harp_consensus_protocol: etcd', explicitly add 'etcd' to
the role for each of the following instances:

-   BDR Primary ('bdr' role)
-   BDR Logical Standby ('bdr' + 'readonly' roles)
-   only for the Bronze layout: BDR Witness ('bdr' + 'witness' roles)
-   only for the Gold layout: Barman ('barman' role)

Note that this is one of the possible choices for the template, and
not necessarily the best one; different placements are possible, as
long as there are three etcd instances for each location, and the BDR
nodes have a local etcd node, which seems convenient.

Previously etcd was placed on the primary and witness nodes, which did
not produce a good template for the Gold and Platinum layouts.

Signed-off-by: Gianni Ciolli <gianni.ciolli@enterprisedb.com>