Teach update-postgres to upgrade from BDR3 to BDR4
Installing BDR4 packages when BDR3+pglogical3 are already installed is
not as simple as "apt-get install bdr4". The new packages cannot coexist
with the old (and specifically not with pglogical3), so we must remove
the latter before installing BDR4.

Also, we must remove pglogical from shared_preload_libraries, so we must
include postgres/config (which already knows how to do that), so we must
delay restart until after the configuration has been updated, which also
means we must include postgres/config/final after the restart.

The bdr4 package repositories must already be configured for this to
work. On a running bdr3 cluster, change tpa_2q_repositories to point to
the bdr4 repositories, run provision+deploy (which will also deploy any
other changes you've made to config.yml), then run update-postgres, and
it'll install the BDR4 packages.

References: TPA-11