Use apt-cache native capability to resolve deps
apt-rdepends doesn't deal with virtual packages.
apt-cache depends has a --recurse option that does full depth recursion.
To limit the output of package references to just Depends and PreDepends
we can filter out unwanted lines.

This avoids the need for a script to parse and handle deps
and the need for apt-rdepends.