Support ddeb files in local-repo
The debug symbols package for postgresql-13-dbgsym has a different
package type to normal - ddeb. To construct the local repo metadata file
we use dpkg-scanpackages which defaults to only scanning the directory
for "deb" files. To cater for "ddeb" files one uses the argument `-t`.
It appears possible to concatenate 2 Packages files together and zip the
result.