Run `harpctl apply` only if the bootstrap config changes (#596)
The way harp configuration files are structured, they loop over
various host groups to generate the desired config. However,
there is no guarantee as to the placement and sequence of the
hosts in a given group during multiple deploys and redeploys.
This means that the order of hosts in the harp configuration
_can_ potentially change across multiple runs resulting in flase
positives. So we add order to the group evaluations to ensure the
sequence of hosts in the rendered configuration file remains the
same unless there is a real change in the cluster.

References: TPA-72

Signed-off-by: Muhammad Haroon <muhammad.haroon@enterprisedb.com>