Derive local-repo directory parent from os_family
Ansible treats Rocky (and AlmaLinux) as part of the RedHat OS family.
To ensure that the initial directory structure for local-repo parent
matches the name which is created when using the `download-packages`
command, derive the directory name from the value for "os_family" in
the platform image. The stored known images dict for aws and docker
now store os_family information where relevant. If the value is unset it
is taken from the "os" value. If the image is custom then we won't try
to assume the os_family and still print out with a message to setup the
repo manually.