Update documentation for air-gapped.md
Note: we have left out instructions for a local-but-networked repo
because tpaexec doesn't currently, natively support disabling OS package
repos to make it truly disconnected.

In disconnected environments it is common for organisations to maintain
their own package repository services that live within the disconnected
environment, with perhaps strict processes for artifact promotion. One
can still use `tpaexec download-packages` to determine what packages
you need to add to this kind of repository. To make use of this repo it
is necessary to disable all other repositories. In config.yml add:

```yaml
cluster_vars:
  yum_repository_list: []
  tpa_2q_repositories: []
```

One would also need to disable any repositories managed by the OS.
TPAexec doesn't currently support this automatically, but it can be
achieved by adding a custom hook in
`"{{ cluster_dir }}/hooks/pre-deploy.yml"`. The exact contents of the
hook will depend on the environment and is beyond the scope of this
document. We hope to address this in a future release.