Split pkg/install into install/upgrade versions
By default, we only ever install packages with `state: present`, so that
they are installed once and left alone thereafter.

However, if you invoke pkg/install with `allow_package_upgrades: yes`
(as update-postgres does, but deploy does not), we set `state: latest`
so that packages get upgraded. This is safe to do only if you stop the
relevant service(s) first and start them afterwards.