Teach update-postgres to stop/update/restart some other services
It can now handle harp/etcd/pgbouncer, according to the instance's role,
by stopping the service, including xxx/pkg with `allow_package_upgrades`
set, and starting the service again afterwards.

We leave out haproxy for now because it (doesn't matter much and)
restarting it during the update will lose (without more configuration)
the stick-table state we went to the trouble to set up for the update in
the first place.