Refrain from resetting permissions on cert files
Do not run security role to reset permissions by default.
If we have not started afresh or have updated the cert on the controller
do not run the security role.

Note that roles/postgres/user/tasks/tls.yml updates the permissions of
certs and isn't linked to sys/ssl. This should be fixed in a future
release.

Without this change, deploy would generate the server key/certificate
owned by root/root, and postgres/user would subsequently set it to
postgres/postgres, which is required for postgres to start, but sys/ssl
would reset the permissions on a subsequent deploy. As a temporary
workaround, since this role is used only to generate the postgres
certificate right now, we avoid the problem temporarily by setting the
permissions to root/root when first generating the certificate, but not
updating them afterwards.