Disable sys/ssl call from sys main
The role sys/ssl was intended to work with the new way of managing
certificate authentication with postgres. That work has yet to be
completed. As it stands activating the role as standard in sys causes a
conflict with the legacy method in postgres/user tls.yml. The legacy
certificate (for the cluster) is created, by tls.yml after the host
certificate in sys/ssl. On a subsequent run sys/ssl detects the legacy
cert and reverts postgres causing a restart. As this is undesirable we
are disabling automatic call to sys/ssl until the work on cert
authentication can be completed.