TPA-95 Use service based SSL certificates
SSL server certificates have been moved to be created per service with
a directory (default) of /etc/tpa/<service>/<hostname>.cert

If sys/ssl/client/server is called with check_result_varname new
caller specific facts will be set to store file locations of certs,
keys, etc. This can be derived from ssl_service, and if not given and
the role assumes it's the default for the cluster.

* ssl_service == not given, var names = `default_ssl_*_filename`
* ssl_service == postgres, var names = `postgres_ssl_*_filename`
* check_result_varname == foo, var names = `foo_ssl_*_filename`

CA generation will also involve copying it to the node. The value of
ssl_ca_file contains the location.

The sys/ssl role is called by sys by default and will generate a CA
if required by either `use_ssl_cert_authentication` or if `selfsigned`
set to false.

Legacy SSL cluster wide cert is detected early in the role `facts`.
(Note: This was not done in minimal-setup as that doesn't appear to know
the name of the cluster.)