TPA-100 Split check for docker so the version is checked before cgroups
Docker info on cgroups is not available on older versions of docker, for
example v18 as distributed with docker.io on Debian 10 (buster).
Therefore, the check errors with a cryptic message.
This change checks for a compatible version first, separately so we can
point users at upgrade docs if they're in this situation.

Also, the docker install docs go to docker desktop, only docker engine
is required for tpaexec which doesn't carry licensing restrictions. So,
the link has been updated to point directly to docker engine community
edition.