TPA-135 fix local-repo define-repository.yml issues
Noticed in testing that the mv of tmpfile into repo file location
could fail if the file had already been moved and would still return
exit code 201. So rewritten to use if;else instead of subshells, added
``&& exit` after mv command so error of failed mv is passed through to
caller.
Also noticed that the checksum was showing as different, due to extra
new-line character difference, so moved the sum calc into the script to
operate on the actual file that would be moved into place. This now
cleanly exits with zero status when no change is needed.