TPA-135 Rocky docker image missing find and xargs for local-repo
In a disconnected environment on the docker image for Rocky the 2
binaries for find and xargs need to be installed from the local-repo
before all yum repos can be disabled.
Adding a prereq install for all platforms of findutils, which is
available on debian/ubuntu as well.
Moved the fact setting of target_repo_dir earlier as this is used by
install-prereq.yml