Add deployment support for pgd-proxy
pgd-proxy is the successor of harp-proxy, for use with BDR v5. (There is
no equivalent to harp-manager, the functionality is absorbed into BDR.)

pgd-proxy can be run on a postgres (bdr) instance, or a standalone
instance with 'pgd-proxy' in role. It runs as pgd_proxy_{user,group}
(per custom service unit definition), and also connects to Postgres as
$pgd_proxy_user (which we create and grant the bdr_superuser role).

Some related pgdcli changes are included alongside. pgdcli needs to be
installed on pgd-proxy standalone instances as well, where its config is
owned by the pgd_proxy_{user,group} instead of postgres_{user,group}. We
also update the contents of the configuration file, and its location.

(With review and major contributions from Jonathan and the TPAexec
developers.)

References: TPA-194