Add support for non harp etcd packages
During patroni development it was observed by PM that projects that
require etcd packages should not be able to use the harp-etcd packages
due to restrictions on products. e.g. non-bdr customers should use
opensource packages from public repos.

Since etcd has been added to PGDG for Redhat and Debian/Ubuntu has its
own etcd packages, there should be an option to use these instead of EDB
hosted packages.

Packages held in public repos for debian and in PGDG for RedHat are of
an older version and supply systemd service unit files. Some adjustments
have been made to the ordering of tasks to ensure configuration is laid
down before the package is installed. If the package is installed first
then an incompatible cluster configuration is created, the data
directory is placed in /var/lib/etcd/default and peers are unable to
connect and the systemd service fails to start.

For harp-etcd it was observed that the daemon runs as root, therefore
a user and group is created for etcd (in all cases as one is added
in the case of patroni packages). For existing clusters running as root
the data directory ownership is changed before the daemon is signaled
to restart. If you want to keep etcd running as root, although this is
not recommended from a security perspective, add this to config.yml:

   cluster_vars:
     etcd_user: root
     etcd_group: root

The `etcd/start` role checks the actual state of the service before
attempting to start. If etcd is in a failed state we don't touch it, but
still allow the `restart` role to attempt a restart.

The `etcd/pkg` role sets a default set of packages that are selected if
the `failover_manager` is not in the list of keys available in
`etcd_packages`. This means that it if `etcd` is in `role` for a node
then the default etcd will be installed, even if the `failover_manager`
is not recognised. This is different to the intended configuration in
the previous revision of code which would not install any packages.

References: TPA-314