Ensure a single etcd cluster for patroni
With the way etcd was setup for harp combined with how patroni was
configured, there would be etcd nodes created in different locations
that did not join the etcd cluster.

Adding etcd_location var disassociates the meaning of the use of this
var as something just for harp. It still works for harp if not set.
Setting it during configure ensures that the etcd cluster can span
multiple locations for HA.

References: TPA-154, PAT-1